requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0

# API (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (optional, for exploration)
jupyter>=1.0.0
ipykernel>=6.25.0

